+++ src/PS.FarNet.Redis/Abc.cs
@@ -1,4 +1,5 @@
 ﻿using StackExchange.Redis;
+using System.Collections;
 using System.Collections.Generic;
 using System.Management.Automation;
 
@@ -6,8 +7,6 @@ namespace PS.FarNet.Redis;
 
 static class Abc
 {
-    public static string[] OneNullString = [null];
-
     public static object BaseObject(object value)
     {
         return value is PSObject ps ? ps.BaseObject : value;
@@ -43,5 +42,40 @@ static class Abc
         foreach (var value in values)
             result.Add((string)value);
         return result;
+    }
+
+    public static RedisValue[] ToRedis(string[] input)
+    {
+        var entries = new RedisValue[input.Length];
+        for (int i = 0; i < entries.Length; ++i)
+            entries[i] = new RedisValue(input[i]);
+
+        return entries;
+    }
+
+    public static HashEntry[] ToRedis(IDictionary input)
+    {
+        var entries = new HashEntry[input.Count];
+        var i = -1;
+        foreach (DictionaryEntry kv in input)
+        {
+            ++i;
+            entries[i] = new(new RedisValue(kv.Key?.ToString()), new RedisValue(kv.Value?.ToString()));
+        }
+
+        return entries;
+    }
+
+    public static KeyValuePair<RedisKey, RedisValue>[] ToRedisPairs(IDictionary input)
+    {
+        var entries = new KeyValuePair<RedisKey, RedisValue>[input.Count];
+        var i = -1;
+        foreach (DictionaryEntry kv in input)
+        {
+            ++i;
+            entries[i] = new(new RedisKey(kv.Key?.ToString()), new RedisValue(kv.Value?.ToString()));
+        }
+
+        return entries;
     }
 }