+++ src/PS.FarNet.Redis/BaseGetCountCmdlet.cs
@@ -4,6 +4,8 @@ namespace PS.FarNet.Redis;
 
 public abstract class BaseGetCountCmdlet : BaseKeyCmdlet
 {
-    [Parameter(ParameterSetName = "Count", Mandatory = true)]
+    protected const string NCount = "Count";
+
+    [Parameter(ParameterSetName = NCount, Mandatory = true)]
     public SwitchParameter Count { get; set; }
 }